Fixed register controller

+++ app/Http/Controllers/MapController.php
@@ -18,26 +18,22 @@ class MapController extends Controller
     {
         //return \App\Models\Map::whereDoesntHave('incidents')->delete();
         $category_ids = $request->input('category_ids');
-        if ($request->is('api*')) {
-            $query = Map::with('categories')->withCount('incidents');
-            if ($request->input('ids')) {
-                $query->whereIn('uuid', $request->input('ids'));
-            } else {
-                $query->where('privacy', 'public');
-            }
+        $query = Map::with('categories')->withCount('incidents');
 
-            $query->when($request->input('category_ids'), function ($query, $category_ids) {
-                return $query->whereHas('categories', function ($q) use ($category_ids) {
-                    $q->whereIn('category_id', $category_ids);
-                });
-            });
-
-            $query->orderBy($request->input('orderBy', 'created_at'), 'desc');
-
-            return $query->get();
+        if ($request->input('ids')) {
+            $query->whereIn('uuid', $request->input('ids'));
         } else {
-            return view('map');
+            $query->where('privacy', 'public');
         }
+
+        $query->when($request->input('category_ids'), function ($query, $category_ids) {
+            return $query->whereHas('categories', function ($q) use ($category_ids) {
+                $q->whereIn('category_id', $category_ids);
+            });
+        });
+        $query->orderBy($request->input('orderBy', 'created_at'), 'desc');
+
+        return $query->get();
     }
 
     /**
@@ -58,7 +54,7 @@ class MapController extends Controller
      */
     public function store(Request $request)
     {
-        // $this->authorize('create', Category::class);
+        $this->authorize('create', Map::class);
         $validatedData = $request->validate([
             'title' => 'nullable|string|max:255',
             'slug' => 'nullable|string|max:255',
@@ -67,7 +63,6 @@ class MapController extends Controller
             'users_can_create_incidents' => 'nullable|in:yes,only_logged_in,no',
         ]);
 
-        // // $image_path = $request->file('icon')->store('categories');
         $uuid = (string) Uuid::generate(4);
         $token = str_random(32);
 
@@ -89,7 +84,7 @@ class MapController extends Controller
         if ($request->is('api*')) {
             return $result;
         } else {
-            return redirect('/maps/'.$result->slug)->with('token', $result->token);
+            return redirect('/maps/' . $result->slug)->with('token', $result->token);
         }
     }
 
@@ -101,14 +96,16 @@ class MapController extends Controller
      */
     public function show(Request $request, Map $map)
     {
-        $data = [
-            'token' => $request->is('api*') ? null : $request->session()->get('token'),
-            'map' => $map->load('categories'),
-        ];
+        $this->authorize('view', $map);
 
         if ($request->is('api*')) {
             return $map;
         }
+
+        $data = [
+            'token' => $request->is('api*') ? null : $request->session()->get('token'),
+            'map' => $map->load('categories'),
+        ];
 
         return View::make('map', $data);
     }
@@ -121,6 +118,8 @@ class MapController extends Controller
      */
     public function showEmbed(Request $request, Map $map)
     {
+        $this->authorize('view', $map);
+
         $data = [
             'map' => $map->load('categories'),
         ];
@@ -148,11 +147,11 @@ class MapController extends Controller
      */
     public function update(Request $request, Map $map)
     {
-        $validatedData = $request->validate([
-            'token' => 'required|exists:maps,token',
+        $this->authorize('update', $map);
 
+        $validatedData = $request->validate([
             'title' => 'nullable|string|max:255',
-            'slug' => 'nullable|string|max:255|unique:maps,slug,'.$map->id,
+            'slug' => 'nullable|string|max:255|unique:maps,slug,' . $map->id,
             'description' => 'nullable|string',
             'privacy' => 'nullable|in:public,unlisted,private',
             'users_can_create_incidents' => 'nullable|in:yes,only_logged_in,no',
@@ -173,7 +172,7 @@ class MapController extends Controller
      */
     public function destroy(Request $request, Map $map)
     {
-        $validatedData = $request->validate(['token' => 'required|exists:maps,token']);
+        $this->authorize('forceDelete', $map);
         $map->delete();
     }
 }
